Write compat_config.xml per class and merge later.
Instead of writing compat_config.xml into the "compat" package, put it in
the java package that defines it including the class name in the filename.
Then merge it later in the build with a python script.

This means that if we get compat_config.xml from static dependencies or
multiple compilation shards, they will be merged rather than overwritten
during jar file merging.

Test: atest compat-changeid-annotation-processor-test
Test: m
Change-Id: Iede6a4040d023820ef735ac611f7f73c0fe0035e