Pretty print metadata xml file to make it easier to read
The process_compat_config.py script can generate two separate XML
files. One for consumption by the device and a separate metadata XML
file that are merged together and used for various tools and tests.

In preparation for including the metadata into the sdk snapshot, which
will be checked into git this change pretty prints the metadata to make
it easier to read and review the differences.

Forces the binary and test to use python3 as ElementTree.indent was
only added in 3.9.

Bug: 232401814
Test: atest --host process-compat-config-test CtsAppCompatHostTestCases
Change-Id: I2c7efc68722eefeca397f7b4d257d5b0c0679920