Resolve compat perf regression via @NoLogging annotation
This annotation allows marking specific compat changes so that they are not logged by the compat framework. This is expected to be used in code that is called in performance critical paths.

Note this change also means that non-loggable changes (including targeting older sdk) are no longer going to stats log.

Change-Id: I7a41ffe70aa924aece9609d44460b3fa67ef00b7
Flag: EXEMPT bugfix
Bug: 424051595